Update 09-ruby.md
codeblocked `before` on line 651
660: capitalisation on Note

"Serve the index.erb view on the root url with the following code block."
WHAT DOES THAT EVEN MEEEANNNNNNN?!?! ;___;

that whole paragraph is kind of confusing/nonsensical to a beginner :(

AND THIS IS DEV-ENVS FOR *BEGINNERS*, RIGHT?! ;_;

yep, you've lost me. well, not really. but the language you're using is pretty dense for a beginner... :(

explain what erb is here