Update 09-ruby.md
line 710: not sure why but it shows up as totally blank in the pdf! :(
really need a proofreader for this ish, ugh >_<;;;

813: should it be `nodemon`?

also what is nodemon?!
isn't that just shotgun?

also...
it totally did not work for me.
lol.
>_<
wait.
it totally did
TELL PEOPLE TO GO TO localhost:9393 afterwards! i was totally confused and also went to localhost:4567 instead ugh :(

it's so beautiful @_@!?